Add sidebar buttons for creating objects
