+++ src/ThreeEditor/js/sidebar/Sidebar.Scene.js
@@ -1,12 +1,9 @@
-import * as THREE from 'three';
-import { UIBreak, UIColor, UIPanel, UIRow, UISelect, UIText } from '../libs/ui.js';
-import { UITexture } from '../libs/ui.three.js';
+import { UIBreak, UIPanel } from '../libs/ui.js';
 import { OutlinerManager } from './Sidebar.OutlinerManager';
-
-
+import { FigureAddPanel, ZoneAddPanel } from './Sidebar.AddPanel';
 
 function SidebarScene(editor) {
-	const { signals, strings } = editor;
+	const { signals } = editor;
 
 	const container = new UIPanel();
 	container.setBorderTop('0');
@@ -27,141 +24,14 @@ function SidebarScene(editor) {
 
 	container.add(new UIBreak());
 
-	// background
+	const figureAddPanel = FigureAddPanel(editor, container);
 
-	const backgroundRow = new UIRow();
-
-	const backgroundType = new UISelect()
-		.setOptions({
-			None: '',
-			Color: 'Color',
-			Texture: 'Texture',
-			Equirectangular: 'Equirect'
-		})
-		.setWidth('160px');
-	backgroundType.onChange(() => {
-		onBackgroundChanged();
-		refreshBackgroundUI();
-	});
-
-	backgroundRow.add(new UIText(strings.getKey('sidebar/scene/background')).setWidth('90px'));
-	backgroundRow.add(backgroundType);
-
-	const backgroundColor = new UIColor()
-		.setValue('#000000')
-		.setMarginLeft('8px')
-		.onInput(onBackgroundChanged);
-	backgroundRow.add(backgroundColor);
-
-	const backgroundTexture = new UITexture().setMarginLeft('8px').onChange(onBackgroundChanged);
-	backgroundTexture.setDisplay('none');
-	backgroundRow.add(backgroundTexture);
-
-	const backgroundEquirectangularTexture = new UITexture()
-		.setMarginLeft('8px')
-		.onChange(onBackgroundChanged);
-	backgroundEquirectangularTexture.setDisplay('none');
-	backgroundRow.add(backgroundEquirectangularTexture);
-
-	container.add(backgroundRow);
-
-	function onBackgroundChanged() {
-		signals.sceneBackgroundChanged.dispatch(
-			backgroundType.getValue(),
-			backgroundColor.getHexValue(),
-			backgroundTexture.getValue(),
-			backgroundEquirectangularTexture.getValue()
-		);
-	}
-
-	function refreshBackgroundUI() {
-		const type = backgroundType.getValue();
-
-		backgroundType.setWidth(type === 'None' ? '160px' : '110px');
-		backgroundColor.setDisplay(type === 'Color' ? '' : 'none');
-		backgroundTexture.setDisplay(type === 'Texture' ? '' : 'none');
-		backgroundEquirectangularTexture.setDisplay(type === 'Equirectangular' ? '' : 'none');
-	}
-
-	// environment
-
-	const environmentRow = new UIRow();
-
-	const environmentType = new UISelect()
-		.setOptions({
-			None: '',
-			Equirectangular: 'Equirect',
-			ModelViewer: 'ModelViewer'
-		})
-		.setWidth('160px');
-	environmentType.setValue('None');
-	environmentType.onChange(() => {
-		onEnvironmentChanged();
-		refreshEnvironmentUI();
-	});
-
-	environmentRow.add(new UIText(strings.getKey('sidebar/scene/environment')).setWidth('90px'));
-	environmentRow.add(environmentType);
-
-	const environmentEquirectangularTexture = new UITexture()
-		.setMarginLeft('8px')
-		.onChange(onEnvironmentChanged);
-	environmentEquirectangularTexture.setDisplay('none');
-	environmentRow.add(environmentEquirectangularTexture);
-
-	container.add(environmentRow);
-
-	function onEnvironmentChanged() {
-		signals.sceneEnvironmentChanged.dispatch(
-			environmentType.getValue(),
-			environmentEquirectangularTexture.getValue()
-		);
-	}
-
-	function refreshEnvironmentUI() {
-		const type = environmentType.getValue();
-
-		environmentType.setWidth(type !== 'Equirectangular' ? '160px' : '110px');
-		environmentEquirectangularTexture.setDisplay(type === 'Equirectangular' ? '' : 'none');
-	}
+	const sceneAddPanel = ZoneAddPanel(editor, container);
 
 	//
 
 	function refreshUI() {
-		const { camera, scene } = editor;
-		const { detectContainer } = editor.detectManager;
-		const { zoneContainer, worldZone } = editor.zoneManager;
-
 		refreshOptions();
-
-		if (scene.background) {
-			if (scene.background.isColor) {
-				backgroundType.setValue('Color');
-				backgroundColor.setHexValue(scene.background.getHex());
-			} else if (scene.background.isTexture) {
-				if (scene.background.mapping === THREE.EquirectangularReflectionMapping) {
-					backgroundType.setValue('Equirectangular');
-					backgroundEquirectangularTexture.setValue(scene.background);
-				} else {
-					backgroundType.setValue('Texture');
-					backgroundTexture.setValue(scene.background);
-				}
-			}
-		} else {
-			backgroundType.setValue('None');
-		}
-
-		if (scene.environment) {
-			if (scene.environment.mapping === THREE.EquirectangularReflectionMapping) {
-				environmentType.setValue('Equirectangular');
-				environmentEquirectangularTexture.setValue(scene.environment);
-			}
-		} else {
-			environmentType.setValue('None');
-		}
-
-		refreshBackgroundUI();
-		refreshEnvironmentUI();
 	}
 
 	refreshUI();
@@ -174,9 +44,7 @@ function SidebarScene(editor) {
 
 	signals.objectChanged.add(refreshUI);
 
-
 	return container;
 }
 
 export { SidebarScene };
-
